% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_customers.R
\name{add_customer}
\alias{add_customer}
\title{Polished Payments API - Add a Customer}
\usage{
add_customer(
  app_uid,
  user_uid,
  stripe_customer_id,
  stripe_subscription_id = NULL,
  is_live = .pp$is_live,
  api_key = polished::get_api_key()
)
}
\arguments{
\item{app_uid}{the app uid.}

\item{user_uid}{the user uid.}

\item{stripe_customer_id}{the Stripe customer id.}

\item{stripe_subscription_id}{the Stripe subscription id.}

\item{is_live}{whether or not the Stripe key is live or test mode.}

\item{api_key}{your Polished API key. Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished Payments API - Add a Customer
}
\seealso{
[get_customers()] [update_customer()]
}
