% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_customer.R
\name{add_stripe_customer}
\alias{add_stripe_customer}
\title{Create a new Stripe customer}
\usage{
add_stripe_customer(email, user_uid)
}
\arguments{
\item{email}{The new Stripe customer's email address.}

\item{user_uid}{the new Stripe customer's polished user uid.}
}
\description{
Create a new Stripe customer
}
