% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_list_to_df}
\alias{api_list_to_df}
\title{Convert a list returned from the API back into a data frame}
\usage{
api_list_to_df(api_list)
}
\arguments{
\item{api_list}{the data frame in list format as returned from the API}
}
\description{
We send data frames back from the API as lists.  This way NA columns are not
dropped when converting to and from JSON.
}
