% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module.R
\name{app_module}
\alias{app_module}
\title{The server logic for the "Admin Panel" 'shinydashboard'}
\usage{
app_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
The 'shiny' module server logic for the Admin Panel.
}
