% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module.R
\name{app_module_ui}
\alias{app_module_ui}
\title{The payments Shiny app}
\usage{
app_module_ui(
  id,
  custom_ui = NULL,
  app_name = getOption("polished")$app_name_display
)
}
\arguments{
\item{id}{the Shiny module id}

\item{custom_ui}{Either \code{NULL}, the default, or a list of 2 elements containing custom
UI to add additional 'shinydashboard' tabs to the 'polished' "Admin Panel".}

\item{app_name}{the app name to display to users.}
}
\value{
the UI for the "Admin Panel"
}
\description{
The Shiny module UI for the Stripe payments Shiny app.  This app can be easily added
to your Shiny apps that use polished.
}
