% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_module.R
\name{billing_module}
\alias{billing_module}
\title{billing module}
\usage{
billing_module(input, output, session, app_url, sub_info)
}
\arguments{
\item{app_url}{the url of the app}

\item{sub_infor}{the subscription information}
}
\description{
billing module
}
