% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_user_subscription.R
\name{check_user_subscription}
\alias{check_user_subscription}
\title{check that the user has a subscription}
\usage{
check_user_subscription(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the Shiny session}
}
\description{
This function should be placed at the top of your Shiny server function
}
\details{
If the user has a subscription that is either in it's free trial period, or the
user has set up a payment method, then allow the user to continue on to the
app.  If they do user does not have a subscription, set them up with the default
subscription.  If the user has a subscription in which the free trial has expired,
but they have not yet enabled a payment method, then redirect them to their
account page where they can set up their subscription.
}
