% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{create_payment_module}
\alias{create_payment_module}
\title{Shiny module server for creating a one time payment}
\usage{
create_payment_module(
  input,
  output,
  session,
  amount,
  currency = "usd",
  send_receipt_email = TRUE,
  description = NULL,
  ui = NULL
)
}
\arguments{
\item{amount}{the amount of the one time payment}

\item{currency}{the currency to use for the payment}

\item{description}{short payment description.  Useful for identifying the payment.}

\item{ui}{optional UI to place above the submit button.  This is often used to include
a disclaimer.}

\item{send_email_receipt}{boolean - whether or not to send an email receipt.  Defaults
to \code{TRUE}.}
}
\value{
a list with 1 reactiveVal
- payment_response - returns the response to the payment attempt
}
\description{
Shiny module server for creating a one time payment
}
