% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{create_payment_module_ui}
\alias{create_payment_module_ui}
\title{Shiny module ui for creating a one time payment}
\usage{
create_payment_module_ui(id)
}
\arguments{
\item{id}{the module id}
}
\description{
Shiny module ui for creating a one time payment
}
