% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_module.R
\name{credit_card_module}
\alias{credit_card_module}
\title{Server logic for Stripe Credit Card input}
\usage{
credit_card_module(
  input,
  output,
  session,
  open_modal_trigger = reactive(0),
  disclaimer_text = "Disclaimer",
  plan_to_enable = NULL,
  subscription = TRUE,
  payment_amount = NULL,
  currency = "usd"
)
}
\arguments{
\item{open_modal_trigger}{The trigger to open the Credit Card modal}

\item{disclaimer_text}{the Disclaimer text}

\item{plan_to_enable}{the subscription plan to enable}

\item{subscription}{whether the Credit Card is for a subscription payment (default) or one time payment}

\item{payment_amount}{the amount of the one time payment}

\item{currency}{the currency of the one time payment}

\item{sub_info}{the Customer's subscription info}
}
\description{
Server logic for Stripe Credit Card input
}
