% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_modal.R
\name{credit_card_subscription_modal_module}
\alias{credit_card_subscription_modal_module}
\title{Credit Card Modal to create a subscription}
\usage{
credit_card_subscription_modal_module(
  input,
  output,
  session,
  price_id,
  title = NULL,
  size = c("m", "s", "l"),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{price_id}{the Stripe price id}

\item{title}{the title to pass to \code{shiny::modalDialog}}

\item{size}{the size to pas to \code{shiny::modalDialog}}

\item{easyClose}{the fade to pass to \code{shiny::modalDialog}}
}
\description{
Credit Card Modal to create a subscription
}
