% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_trial_banner_module.R
\name{free_trial_banner_module}
\alias{free_trial_banner_module}
\title{Free trial banner module server logic}
\usage{
free_trial_banner_module(input, output, session)
}
\arguments{
\item{input}{the Shiny input}

\item{output}{the Shiny output}

\item{session}{the Shiny session}
}
\description{
Free trial banner module server logic
}
