% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_trial_banner_module.R
\name{free_trial_banner_module_ui}
\alias{free_trial_banner_module_ui}
\title{free trial banner}
\usage{
free_trial_banner_module_ui(id)
}
\arguments{
\item{id}{the Shiny module id}
}
\description{
A banner to show to users that are in their free trial to let them know they
will need to enable billing to continue using the app after their free trial
has expired.
}
