% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_customers.R
\name{get_customers}
\alias{get_customers}
\title{Polished Payments API - Get Customer(s)}
\usage{
get_customers(
  app_uid = NULL,
  user_uid = NULL,
  is_live = getOption("pp")$is_live,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{an optional app uid.}

\item{user_uid}{an optional user uid.}

\item{is_live}{whether or not the Stripe key is live or test mode.}

\item{api_key}{your Polished API key. Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  The "content" of the object is a
tibble of app(s) with the following columns:
- uid
- app_name
- app_url
- created_at
- modified_at
}
\description{
Polished Payments API - Get Customer(s)
}
\details{
If both the \code{app_uid} and \code{user_uid} are \code{NULL}, then all the
customers in your account will be returned.  If only \code{app_uid} is not \code{NULL},
then all customers for a single app will be returned.  If only \code{user_uid} is not
\code{NULL} then all customers for that user will be returned (a user can have multiple
customers for multiple apps).  If \code{app_uid} and \code{user_uid} are provided, then
one customer will be returned (assuming that customer exists). If the customer does not
exists, a zero row tibble will be returned.
}
\seealso{
[add_customer()] [update_customer()]
}
