% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_polished_payments.R
\name{get_polished_payments}
\alias{get_polished_payments}
\title{Add the polished session data to your user}
\usage{
get_polished_payments(polished_api_key, session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{the Shiny session}
}
\description{
Calling this function in your Shiny server will attached your user's Polished
payments data to a session$userData$payments() reactiveVal.
}
