% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_customer.R
\name{get_stripe}
\alias{get_stripe}
\title{Get relevant information about a Stripe customer}
\usage{
get_stripe(
  user_uid,
  user_roles,
  is_on_payments = FALSE,
  app_uid = getOption("polished")$app_uid,
  stripe_prices = getOption("pp")$prices,
  free_roles = getOption("pp")$free_roles
)
}
\arguments{
\item{user_uid}{the customer's `polished` user uid}

\item{user_roles}{the customer's `polished` user roles}

\item{is_on_payments}{whether the customer is on the payments page or not (Default: \code{FALSE})}

\item{app_uid}{the `polished` app uid; defaults to the current app
(e.g. \code{getOption("polished")$app_uid})}

\item{stripe_prices}{the Stripe prices; defaults to the current Stripe prices
provided in \code{polished_payments_config} (e.g. \code{getOption("pp")$prices})}

\item{free_roles}{the roles for free Stripe users; defaults to the roles
provided in \code{polished_payments_config} (e.g. \code{getOption("pp")$free_roles})}
}
\description{
Get relevant information about a Stripe customer
}
