% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_payment_method.R
\name{get_stripe_payment_method}
\alias{get_stripe_payment_method}
\title{get default payment method from Stripe}
\usage{
get_stripe_payment_method(payment_method_id)
}
\arguments{
\item{payment_method_id}{the Stripe payment method id.}
}
\description{
Retrieved a list of information about the polishedpayments customer's default
payment method.
}
