% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_to_account.R
\name{go_to_account}
\alias{go_to_account}
\title{go_to_account}
\usage{
go_to_account(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A Shiny session object.}
}
\description{
trigger event to redirect user to Account (payments) page,
typically used in an `observeEvent`
}
