% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{pay_button_module}
\alias{pay_button_module}
\title{Server logic to create a Stripe credit card payment}
\usage{
pay_button_module(
  input,
  output,
  session,
  payment_amount = NULL,
  currency = "usd"
)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{shiny server session}

\item{payment_amount}{the amount of the one time payment}

\item{currency}{the currency of the one time payment}

\item{open_modal_trigger}{The trigger to open the Credit Card modal}

\item{disclaimer_text}{the Disclaimer text}
}
\description{
Server logic to create a Stripe credit card payment
}
