% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_app.R
\name{payments_app_server}
\alias{payments_app_server}
\title{The server logic for the "Admin Panel" 'shinydashboard'}
\usage{
payments_app_server(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
The 'shiny' module server logic for the Admin Panel.
}
