% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_app.R
\name{payments_app_ui}
\alias{payments_app_ui}
\title{The payments Shiny app}
\usage{
payments_app_ui(app_name = polished::.polished$app_name)
}
\arguments{
\item{app_name}{the app name to display to users.}
}
\value{
the UI for the "Admin Panel"
}
\description{
The Shiny module UI for the Stripe payments Shiny app.  This app can be easily added
to your Shiny apps that use polished.
}
