% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_server.R
\name{payments_server}
\alias{payments_server}
\title{adds polished payments to your Shiny server}
\usage{
payments_server(server)
}
\arguments{
\item{server}{the Shiny server function}
}
\description{
Wrap your Shiny server in this function to add polished payments to your Shiny
server before your custom Shiny app's server logic runs.
}
