% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments_ui.R
\name{payments_ui}
\alias{payments_ui}
\title{adds polished payments to your Shiny UI}
\usage{
payments_ui(ui, app_name = getOption("polished")$app_name)
}
\arguments{
\item{ui}{the Shiny app UI}

\item{app_name}{the app name to display to the user in the return to app link.}
}
\description{
adds polished payments to your Shiny UI
}
