% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_payments_config.R
\name{polished_payments_config}
\alias{polished_payments_config}
\title{configure Polished Payments}
\usage{
polished_payments_config(
  stripe_secret_key,
  stripe_public_key,
  subscription_prices = NULL,
  trial_period_days = 0,
  is_subscription_required = TRUE
)
}
\arguments{
\item{stripe_secret_key}{the Stripe secret key}

\item{stripe_public_key}{the Stripe publishable key}

\item{subscription_prices}{an unnamed character vector of Stripe price ids for your subscription.  e.g. monthly and
yearly pricing options.  A Stripe price id looks like this "price_64t6gq76vr78sdhf".  Set to \code{NULL}, the default,
if you are only using one time payments.}

\item{trial_period_days}{the number of days to offer for a free trial period.  All pricing options
will use this free trial period.  It overrides any free trial period set on your Stripe dashboard.}

\item{is_subscription_required}{\code{TRUE} or \code{FALSE} for whether or not a subscription is
required.  If there are no subscription prices passed to \code{subscription_prices} argument, then
this argument is always treated as \code{FALSE}.}
}
\description{
configure Polished Payments
}
\examples{

\dontrun{
polished_payments_config(
  stripe_secret_key = "<your Stripe secret API key>",
  stripe_public_key = "<your Stripe publishable key>",
  subscription_prices = c("price_jkashdkfjh", "price_jakhkljgakwf"),
  trial_period_days = 30
)
}


}
