% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_payment_method_modal.R
\name{set_payment_method_modal}
\alias{set_payment_method_modal}
\title{Modal to set/update the Stripe payment method}
\usage{
set_payment_method_modal(
  input,
  output,
  session,
  open_modal_trigger,
  title = "Enable Billing"
)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{open_modal_trigger}{the reactive trigger to open the modal}

\item{title}{the title of the modal}
}
\description{
Modal to set/update the Stripe payment method
}
