% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_customers.R
\name{update_customer}
\alias{update_customer}
\title{Polished Payments API - Update a Customer}
\usage{
update_customer(
  customer_uid,
  stripe_subscription_id = NULL,
  free_trial_days_remaining_at_cancel = NULL,
  default_payment_method = NULL,
  stripe_is_live = getOption("pp")$is_live,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{customer_uid}{The polished payments customer UID.}

\item{stripe_subscription_id}{the Stripe subscription ID.}

\item{free_trial_days_remaining_at_cancel}{The number of days remaining in the customer's
free subscription trial at time of cancelation.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished Payments API - Update a Customer
}
\seealso{
[get_customers()] [add_customer()]
}
