% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_module.R
\name{billing_module}
\alias{billing_module}
\title{billing module}
\usage{
billing_module(input, output, session, sub_info)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{sub_info}{the subscription information}
}
\description{
billing module
}
