% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_payment_method.R
\name{create_payment}
\alias{create_payment}
\title{create a Stripe payment}
\usage{
create_payment(
  customer_id,
  payment_method_id,
  amount,
  currency = "usd",
  receipt_email = NULL,
  description = NULL,
  stripe_secret_key = getOption("pp")$keys$secret
)
}
\arguments{
\item{customer_id}{the Stripe customer ID}

\item{payment_method_id}{the Stripe payment method ID}

\item{amount}{a positive integer representing how much to charge
in the smallest currency unit (e.g., 100 cents to charge $1.00 or 100 to charge ¥100, a zero-decimal currency)}

\item{currency}{the currency}

\item{receipt_email}{the email address to receive the receipt}

\item{description}{payment description to show to user making payment}

\item{stripe_secret_key}{your Stripe account secret API key}
}
\description{
create a Stripe payment
}
