% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{create_payment_module}
\alias{create_payment_module}
\title{Shiny module server for creating a one time payment}
\usage{
create_payment_module(
  input,
  output,
  session,
  amount,
  currency = "usd",
  send_receipt_email = TRUE,
  description = NULL,
  ui = NULL
)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{amount}{a positive integer representing how much to charge in the
smallest currency unit (e.g., 100 cents to charge $1.00 or 100 to charge
¥100, a zero-decimal currency)}

\item{currency}{the currency to use for the payment}

\item{send_receipt_email}{boolean - whether or not to send an email receipt.  Defaults
to \code{TRUE}.}

\item{description}{short payment description.  Useful for identifying the payment.}

\item{ui}{optional UI to place above the submit button.  This is often used to include
a disclaimer.}
}
\value{
a list with 1 reactiveVal
- payment_response - returns the response to the payment attempt
}
\description{
Shiny module server for creating a one time payment
}
