% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{create_payment_module}
\alias{create_payment_module}
\title{Shiny module server for creating a one time payment}
\usage{
create_payment_module(
  input,
  output,
  session,
  amount,
  currency = "usd",
  send_receipt_email = TRUE,
  description = NULL,
  ui = NULL,
  title = NULL,
  size = "s",
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{amount}{the amount of the one time payment}

\item{title}{the title to pass to \code{shiny::modalDialog}}

\item{size}{the size to pas to \code{shiny::modalDialog}}

\item{easyClose}{the easyClose to pass to \code{shiny::modalDialog}}

\item{fade}{the fade to pass to \code{shiny::modalDialog}}

\item{open_modal_trigger}{reactive trigger to open the modal}
}
\description{
Shiny module server for creating a one time payment
}
