% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_module.R
\name{create_payment_module_ui}
\alias{create_payment_module_ui}
\title{UI element to create a Stripe Credit Card payment}
\usage{
create_payment_module_ui(id)
}
\arguments{
\item{id}{The element's namespaced ID}
}
\description{
UI element to create a Stripe Credit Card payment
}
