% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subscription_modal.R
\name{create_subscription_modal}
\alias{create_subscription_modal}
\title{Credit Card Modal to create a subscription}
\usage{
create_subscription_modal(
  input,
  output,
  session,
  open_modal_trigger,
  price_id,
  ui = tags$p(style = "text-align: center", "You are signing up for a subscription"),
  title = NULL,
  size = "s",
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{open_modal_trigger}{the reactive trigger to open the modal}

\item{price_id}{the Stripe price id}

\item{ui}{custom UI added to bottom of the modal}

\item{title}{the title to pass to \code{shiny::modalDialog}}

\item{size}{the size to pas to \code{shiny::modalDialog}}

\item{easyClose}{the easyClose to pass to \code{shiny::modalDialog}}

\item{fade}{the fade to pass to \code{shiny::modalDialog}}
}
\description{
Credit Card Modal to create a subscription
}
