% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_module.R
\name{credit_card_module}
\alias{credit_card_module}
\title{Server logic for Stripe Credit Card input}
\usage{
credit_card_module(
  input,
  output,
  session,
  trigger = function() NULL,
  billing_details = function() list()
)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{trigger}{the reactive trigger to submit the payment to Stripe}

\item{billing_details}{reactive returning a named list of billing details. Valid list element names
are:
- "name"
- "email"
- "phone"}
}
\description{
Server logic for Stripe Credit Card input
}
