% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_module.R
\name{credit_card_module_ui}
\alias{credit_card_module_ui}
\title{UI element for Stripe Credit Card input}
\usage{
credit_card_module_ui(id)
}
\arguments{
\item{id}{The element's namespaced ID}
}
\description{
UI element for Stripe Credit Card input
}
