% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card_module.R
\name{credit_card_payment_module}
\alias{credit_card_payment_module}
\title{Server logic for Stripe Credit Card input}
\usage{
credit_card_payment_module(
  input,
  output,
  session,
  trigger = function() NULL,
  amount,
  currency = "usd"
)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{trigger}{the reactive trigger to submit the payment to Stripe}

\item{amount}{the amount of the payment}

\item{currency}{The currency. Defaults to "usd" (United States Dollar).}
}
\description{
Server logic for Stripe Credit Card input
}
