% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stripe.R
\name{get_stripe}
\alias{get_stripe}
\title{Get relevant information about a Stripe customer from Stripe and Polished}
\usage{
get_stripe(user_uid, app_uid)
}
\arguments{
\item{user_uid}{the customer's `polished` user uid}

\item{app_uid}{the `polished` app uid; defaults to the current app
(e.g. \code{.polished$app_uid})}
}
\value{
a list with the following elements:
- polished_customer_uid
- stripe_customer_id
- free_user
- default_payment_method
- trial_days_remaining
- subscription: either NA if no subscription in Stripe, or a list of with the following
elements retrieved from the Stripe API:
  - stripe_subscription_id
  - item_id
  - plan_id
  - nickname
  - amount
  - interval
  - trial_end
  - created_at
}
\description{
This is a helper API wrapper for internal use in Polished Payments Shiny apps.
It is not intended to be used outside of Shiny.
}
