% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_customer.R
\name{get_stripe_customer}
\alias{get_stripe_customer}
\title{Collect information for a Stripe customer}
\usage{
get_stripe_customer(
  stripe_customer_id,
  stripe_secret_key = getOption("pp")$keys$secret
)
}
\arguments{
\item{stripe_customer_id}{The Stripe customer's ID.}
}
\description{
Collect information for a Stripe customer
}
