% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_payment_method.R
\name{get_stripe_payment_method}
\alias{get_stripe_payment_method}
\title{get payment method information from Stripe for a particular Payment Method ID}
\usage{
get_stripe_payment_method(payment_method_id)
}
\arguments{
\item{payment_method_id}{the Stripe payment method id.}
}
\description{
Retrieves a list of information about the `polishedpayment` customer's payment method.
}
