% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stripe.R
\name{get_stripe_subscription}
\alias{get_stripe_subscription}
\title{Get information on the user's Stripe subscription}
\usage{
get_stripe_subscription(stripe_subscription_id)
}
\arguments{
\item{stripe_subscription_id}{Your user's Stripe subscription ID.}
}
\description{
Get information on the user's Stripe subscription
}
