% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_payments_config.R
\name{polished_payments_config}
\alias{polished_payments_config}
\title{configure R options for Polished Payments}
\usage{
polished_payments_config(.options)
}
\arguments{
\item{.options}{a list containing your Stripe configuration.  The list should be
in the following format:
list(
  keys = list(
    secret: <your Stripe secret key>
    public: <your Stripe public key>
  ),
  prices = c(<character vector of 1 to 4 pricing options>)
  free_trial_days = <number of days for your free trial>
)}
}
\description{
configure R options for Polished Payments
}
\examples{

polished_payments_config(
  list(
    keys = list(
      secret: "fake_key_hjkasdhgkjashdkfj"
      public: "fake_key_hjaksdfhjaksldhfkj"
    ),
    prices = c("price_jkashdkfjh", "price_jakhkljgakwf")
    free_trial_days = 30
  )
)


}
