% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolishedPayments.R
\name{polished_payments_config}
\alias{polished_payments_config}
\title{configure Polished Payments}
\usage{
polished_payments_config(
  stripe_secret_key,
  stripe_public_key,
  subscription_prices = NULL,
  trial_period_days = 0,
  free_roles = character(0)
)
}
\arguments{
\item{stripe_secret_key}{the Stripe secret key}

\item{stripe_public_key}{the Stripe publishable key}

\item{subscription_prices}{an unnamed character vector of Stripe price ids for your subscription.  e.g. monthly and
yearly pricing options.  A Stripe price id looks like this "price_64t6gq76vr78sdhf".  Set to \code{NULL}, the default,
if you are only using one time payments.}

\item{trial_period_days}{the number of days to offer for a free trial period.  All pricing options
will use this free trial period.  It overrides any free trial period set on your Stripe dashboard.}

\item{free_roles}{Polished user roles that can bypass having to set up a subscription and get free
access to your Shiny app.  This is often used to give certain users (e.g. your beta testers) free
access to your app.  Go to \url{https://dashboard.polished.tech} to create a user role
and add that role to specific users, or use the `polished` package's API wrapper functions
(`polished::create_role` & `polished::add_user_role`).}
}
\description{
configure Polished Payments
}
\examples{

\dontrun{
polished_payments_config(
  stripe_secret_key = "<your Stripe secret API key>",
  stripe_public_key = "<your Stripe publishable key>",
  subscription_prices = c("price_jkashdkfjh", "price_jakhkljgakwf"),
  trial_period_days = 30,
  free_roles = "free_user"
)
}


}
