% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_payments_config.R
\name{polished_payments_config}
\alias{polished_payments_config}
\title{configure R options for Polished Payments}
\usage{
polished_payments_config(
  stripe_secret_key,
  stripe_public_key,
  stripe_prices,
  free_trial_period = 0,
  free_roles = character(0)
)
}
\arguments{
\item{stripe_secret_key}{the Stripe secret key}

\item{stripe_public_key}{the Stripe publishable key}

\item{free_roles}{Polished user roles that can bypass having to set up a aubscription and get free
access to your Shiny app.  This is often used to give certain users (e.g. your beta testers) free
access to your app.  Go to \url{https://dashboard.polished.tech} to create a user role and add that
role to specific users.}

\item{prices}{a character vector of Stripe prices for your subscription.  e.g. monthly and
yearly pricing options.}

\item{free_trial_days}{the number of days to offer for a free trial period.  All pricing options
will use this free trial period.  It overrides any free trial period set on your Stripe dashboard.}
}
\description{
configure R options for Polished Payments
}
\examples{

polished_payments_config(
  stripe_secret_key = "<your Stripe secret API key>",
  stripe_public_key = "<your Stripe publishable key>",
  prices = c("price_jkashdkfjh", "price_jakhkljgakwf"),
  free_trial_days = 30,
  free_roles = "free_user"
)



}
