% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_stripe_payment_method.R
\name{set_stripe_payment_method}
\alias{set_stripe_payment_method}
\title{Set the Default Payment Method for a Customer}
\usage{
set_stripe_payment_method(customer_id, payment_method_id)
}
\description{
Set the Default Payment Method for a Customer
}
